Rename `_gate.py` to `gate` package and update imports
