+++ swanlab/sdk/internal/settings/__init__.py
@@ -38,9 +38,9 @@ from swanlab.proto.swanlab.settings.probe.v1.probe_pb2 import ProbeSettings as P
 from swanlab.sdk.internal.pkg import console, helper, nrc, safe
 from swanlab.sdk.typings.run import ModeType
 
-from ._gate import env_or, load_external_enabled, settings_local
 from .core import CoreSettings
 from .experiment import ExperimentSettings, ProjectSettings, RunSettings
+from .gate import env_or, load_external_enabled, settings_local
 from .integration import IntegrationSettings
 from .probe import ProbeSettings
 from .terminal import TerminalSettings
