+++ swanlab/sdk/internal/settings/gate/__init__.py
@@ -11,6 +11,8 @@
 
 由于各子模块（core / experiment / integration 等）的向下兼容环境变量绕过了pydantic的加载机制，
 因此读取统一通过 env_or 接入门控，保证降级实例为不含任何外部配置的纯默认值。
+
+作为工具模块，为了区分于配置模块，改用文件夹申明模块
 """
 
 import os
